storage refactor to get ready for node, added env key
