+++ packages/client/src/authState/updateAuthState.ts
@@ -1,6 +1,6 @@
 import { AuthState } from '../types'
 import { BasedClient } from '..'
-import { setStorage, removeStorage } from '../localStorage'
+import { setStorage, removeStorage } from '../persistentStorage'
 
 export const updateAuthState = (client: BasedClient, authState: AuthState) => {
   if (authState.persistent) {
